+++ src/hawk/scout/trainer/yolo_radar/trainer.py
@@ -59,11 +59,10 @@ class YOLOTrainerRadar(ModelTrainerBase):
         if version == -1:
             version = self.get_new_version()
 
-        if self.args["mode"] != "oracle":
-            if path is None or not path.is_file():
-                assert len(content)
-                path = self.context.model_path(version, template="model-{}.pt")
-                path.write_bytes(content)
+        if self.args["mode"] != "oracle" and (path is None or not path.is_file()):
+            assert len(content)
+            path = self.context.model_path(version, template="model-{}.pt")
+            path.write_bytes(content)
 
         assert path is not None
 
@@ -148,10 +147,7 @@ class YOLOTrainerRadar(ModelTrainerBase):
         with open(data_file, "w") as outfile:
             yaml.dump(data_dict, outfile, default_flow_style=False)
 
-        if self.train_initial_model:
-            weights = "yolov5s.pt"
-        else:
-            weights = str(self.prev_path)
+        weights = Path("yolov5s.pt") if self.train_initial_model else self.prev_path
 
         cmd = [
             sys.executable,
